Fixed null value bug in value converters
If a value hasn't been specified, the intermediate value sometimes comes back as null rather than an empty string. This may be when the property editor is used in the block list, but not as a regular property on a content type - but I haven't tested enough to confirm this.

The value converters will try to set a fallback value when specified on the data type, but they would return an empty string before reaching the fallback logic if the intermediate value was null. With this commit, the null check has been (re)moved, ensuring that the fallback value is correctly applied.